skeleton for benchmarking workflow (WIP)
